Move Control class to control module
